api implementation

+++ lib/src/features/weather/domain/usecases/current/get_current_weather_usecase.dart
@@ -2,5 +2,5 @@ import 'package:dartz/dartz.dart';
 import 'package:rock_weather/src/features/weather/domain/entities/weather_entity.dart';
 
 abstract class GetCurrentWeatherUsecase {
-  Future<Either<Exception, WeatherEntity>> call(int latitude, int longitude);
+  Future<Either<Exception, WeatherEntity>> call(double latitude, double longitude);
 }
